pyproject.toml
[project]
name = "avthrift-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.7",
    "django-cors-headers>=4.9.0",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "gunicorn>=23.0.0",
    "psycopg[binary]>=3.2.12",
    "PyMySQL>=1.1.0",
    "cryptography>=43.0.0",
    "python-decouple>=3.8",
    "sentry-sdk>=2.43.0",
    "whitenoise>=6.11.0",
    "httpx>=0.27.0",
]

[tool.black]
line-length = 120
target-version = ["py313"]
skip-string-normalization = false

[tool.isort]
profile = "black"
line_length = 120
py_version = 313
src_paths = ["config"]

[tool.flake8]
max_line_length = 120
extend_ignore = ["E203", "W503"]
exclude = [
  ".venv",
  "build",
  "dist",
  "migrations",
  "staticfiles",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "factory-boy>=3.3.3",
    "flake8>=7.3.0",
    "isort>=7.0.0",
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "pytest-django>=4.11.1",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.test"
python_files = ["test_*.py", "*_test.py"]
addopts = "-q --disable-warnings --cov=orders --cov-report=term-missing --cov-fail-under=95"

[tool.coverage.run]
branch = true
source = ["orders"]
omit = [
  "orders/migrations/*",
]

[tool.coverage.report]
show_missing = true
skip_covered = false
